Extract waitUntilReminderHasStatus in wait.js
